feat: Improvement of useRegister + creation of useLogIn
