+++ .env.example
@@ -1,0 +1,6 @@
+VITE_APP_API_KEY=xxx
+VITE_APP_AUTH_DOMAIN=xxx
+VITE_APP_PROJECT_ID=xxx
+VITE_APP_STORAGE_BUCKET=xxx
+VITE_APP_MESSAGING_SENDER_ID=xxx
+VITE_APP_APP_ID=xxx
