+++ README.md
@@ -1,33 +1,41 @@
-# Lingi-club 🗣️
-
-## Introduction 📚
-
-A platform that connects foreign families living in France (often English, German, etc.) with French families who are looking to improve their foreign language skills.
-<br/>
-The foreign family welcomes for a minimum of one week (in batches of weeks) a member or more of a French family who will live the immersion necessary to learn a foreign language without leaving France.
+# Front Folder 📁
 
 ## Getting Started 🚀
 
 To get started with the project, follow these steps in your terminal:
 
-1. **Clone the Repository**  
-   `git clone https://github.com/Caro-L-dev/lingi-club.git`
+1. **Navigate to the Project Directory**  
+   `cd front`
 
-2. **Navigate to the Project Directory**  
-   `cd lingi-club`
+2. **Install Dependencies**  
+   `npm install`
 
-## Contributing 🤝
+3. **Start the Development Server**  
+   `npm run dev`
 
-If you'd like to contribute, please follow these steps:
+The application will now be running in development mode.
+<br/>
+You can access it by navigating to `http://localhost:5173` in your web browser.
 
-1. **Create a New Branch**: Always create a new branch from `develop` to test new features. Use a naming convention that clearly describes the feature or fix you're working on. For example, if you're adding a new calendar for registration feature, you might name your branch `feature/registration-calendar`.
+## Development Workflow 🛠️
 
-2. **Keep `develop` Up to Date**: Before starting work on a new feature or fix, make sure your local `develop` branch is up to date with the remote repository. You can do this by running:  
-   `git checkout develop`  
-   `git pull`
+- **Development Mode**: Run `npm run dev` to start the development server.
+- **Testing**: Run `npm run test` to execute tests using Vitest and React Testing Library.
+- **Building for Production**: Run `npm run build` to create an optimized build of the application.
 
-3. **Submit a Pull Request**: Once you've made your changes and tested them, submit a pull request to the `develop` branch.
+## Testing 🧪
 
----
+The project uses Jest and React Testing Library for testing. To run the tests, execute:  
+`npm run test`
 
-🚀 Happy coding!
+## Tools 🧰
+
+- Vite / React / Typescript / Tailwind
+  <br/>
+- react-router-dom / react-hook-form
+  <br/>
+- shadcn-ui / clsx / lucide-react / react-toastify
+  <br/>
+- uuid / react-query
+  <br/>
+- vitest / react-testing-library
