--- back/README.md
@@ -1,11 +1,0 @@
-# Back Folder 📁
-
-## Getting Started 🚀
-
-To get started with the project, follow these steps in your terminal:
-
-1. **Navigate to the Project Directory**  
-   `cd back`
-
-2. **Start the Development Server**  
-   `?`
