--- front/.env.example
@@ -1,6 +1,0 @@
-VITE_APP_API_KEY=AIzaSyDLxUk1K2HPfbsyj41cGGSGZ_LFxrMa3SM
-VITE_APP_AUTH_DOMAIN=lingui-club-370fe.firebaseapp.com
-VITE_APP_PROJECT_ID=lingui-club-370fe
-VITE_APP_STORAGE_BUCKET=lingui-club-370fe.appspot.com
-VITE_APP_MESSAGING_SENDER_ID=637220579012
-VITE_APP_APP_ID=1:637220579012:web:342e615765e3dba0620b91
