--- front/README.md
@@ -1,41 +1,0 @@
-# Front Folder 📁
-
-## Getting Started 🚀
-
-To get started with the project, follow these steps in your terminal:
-
-1. **Navigate to the Project Directory**  
-   `cd front`
-
-2. **Install Dependencies**  
-   `npm install`
-
-3. **Start the Development Server**  
-   `npm run dev`
-
-The application will now be running in development mode.
-<br/>
-You can access it by navigating to `http://localhost:5173` in your web browser.
-
-## Development Workflow 🛠️
-
-- **Development Mode**: Run `npm run dev` to start the development server.
-- **Testing**: Run `npm run test` to execute tests using Vitest and React Testing Library.
-- **Building for Production**: Run `npm run build` to create an optimized build of the application.
-
-## Testing 🧪
-
-The project uses Jest and React Testing Library for testing. To run the tests, execute:  
-`npm run test`
-
-## Tools 🧰
-
-- Vite / React / Typescript / Tailwind
-  <br/>
-- react-router-dom / react-hook-form
-  <br/>
-- shadcn-ui / clsx / lucide-react / react-toastify
-  <br/>
-- uuid / react-query
-  <br/>
-- vitest / react-testing-library
